requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.10
alembic==1.12.1
pydantic==2.9.2
pydantic-settings==2.4.0
google-cloud-aiplatform==1.38.1
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
email-validator==2.1.0.post1
